All Updated Work And Relocations
All .txt files have been moved to a new file(All Text Descriptions)
New .html files were added for the web app to interact with one another, not yet done need to work on weather and feedback info's.
The .css File has all the styles for the html files already ,all .html files styles are in the same css file and it has comments to make it easier to understand what's happening.